% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_laureate.R
\name{get_laureate}
\alias{get_laureate}
\title{get raw data of nobel prize laureate}
\usage{
get_laureate()
}
\value{
a dataframe of raw data of nobel prize laureate API.
}
\description{
get raw dataframe from API:"http://api.nobelprize.org/v1/laureate.json"
}
\examples{
get_laureate()
}
