% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nobel_country.R
\name{get_nobel_country}
\alias{get_nobel_country}
\title{get country rank information of Nobel Prize winner}
\usage{
get_nobel_country(y = NULL, c = NULL)
}
\arguments{
\item{y}{The year of wining in character.}

\item{c}{The category of the prize in character.}
}
\value{
a dataframe of the rank of country win the Prize given year or category.
}
\description{
This function tells the born country of the winner given year or category.
}
\examples{
get_nobel_country(y = "2018")
get_nobel_country(c = "physics")
}
