% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nobel_gender.R
\name{get_nobel_gender}
\alias{get_nobel_gender}
\title{get gender information of nobel prize}
\usage{
get_nobel_gender(y = NULL, c = NULL, org = TRUE)
}
\arguments{
\item{y}{The year of wining in character.}

\item{c}{The category of the prize in character.}

\item{org}{Choose whether display organization in the output. Input TRUE or FALSE. Default by TRUE.}
}
\value{
a dataframe of gender information of nobel prize.
}
\description{
This function tells the gender(female,male or orgnization) about the Nobel Prize winner given the year or category.
}
\examples{
get_nobel_gender(org = FALSE)
get_nobel_gender(y = "1966")
get_nobel_gender(c = "physics")
}
