% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nobel_info.R
\name{get_nobel_info}
\alias{get_nobel_info}
\title{get basic information of nobel prize}
\usage{
get_nobel_info(y = NULL, c = NULL)
}
\arguments{
\item{y}{The year of wining in character.}

\item{c}{The category of the prize in character.}
}
\value{
a dataframe of some basic information of nobel prize.
}
\description{
This function tells some basic information about the Nobel Prize winner given the year or category.
}
\examples{
get_nobel_info()
get_nobel_info(y = "2018")
get_nobel_info(y = "2018", c = "physics")
}
