% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nobel_share.R
\name{get_nobel_share}
\alias{get_nobel_share}
\title{get share information of Nobel Prize}
\usage{
get_nobel_share(y = NULL, c = NULL)
}
\arguments{
\item{y}{The year in character.}

\item{c}{The category of the prize in character.}
}
\value{
a dataframe of share information of the prize with year and category.
}
\description{
This function tells the share of the prize given year or category.
}
\examples{
get_nobel_age(y = "2018")
get_nobel_age(y = "2018", c = "physics")
}
