% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prize.R
\name{get_prize}
\alias{get_prize}
\title{get raw data of nobel prize}
\usage{
get_prize()
}
\value{
a dataframe of raw data of nobel prize API.
}
\description{
get raw dataframe from API:"http://api.nobelprize.org/v1/prize.json"
}
\examples{
get_prize()
}
